Extracted styles into its own file
