+++ pong_frontend/src/components/mainPages/HomePage.tsx
@@ -1,22 +1,12 @@
 import React from 'react';
 import { useUserContext } from '../context/UserContext';
-import CSS from 'csstype';
-
-const welcomeTitleStyle: CSS.Properties = {
-	color: 'rgba(254, 8, 16, 1)',
-	position: 'relative',
-	textAlign: 'center',
-	top: '8px',
-	padding: '4px',
-	fontFamily: 'Shlop',
-	fontSize: '40px',
-  };
+import * as styles from './HomePageStyles';
 
 const HomePage: React.FC = () => {
 	const { user } = useUserContext();
     return (<div>
-				{ user ? (<p style={welcomeTitleStyle}>Welcome {user.username} {user.userID} to the Horror Ping Pong</p>)
-					   : (<p style={welcomeTitleStyle}>Intruder detected</p>)
+				{ user ? (<p style={styles.welcomeTitleStyle}>Welcome {user.username} {user.userID} to the Horror Ping Pong</p>)
+					   : (<p style={styles.welcomeTitleStyle}>Intruder detected</p>)
 				}
             </div>)
 };
