+++ pong_frontend/src/components/mainPages/HomePageStyles.ts
@@ -1,0 +1,11 @@
+import { Properties } from 'csstype';
+
+export const welcomeTitleStyle: Properties = {
+	color: 'rgba(254, 8, 16, 1)',
+	position: 'relative',
+	textAlign: 'center',
+	top: '8px',
+	padding: '4px',
+	fontFamily: 'Shlop',
+	fontSize: '40px',
+};
