+++ pong_frontend/src/components/mainPages/SettingsPageStyles.ts
@@ -1,0 +1,69 @@
+import { Properties } from 'csstype';
+
+export const profilePictureStyle: Properties = {
+    width: '120px',
+    height: '120px',
+    borderRadius: '50%',
+    objectFit: 'cover',
+    marginBottom: '20px',
+    border: '3px solid rgba(254, 8, 16, 1)',
+};
+
+export const settingsTitleStyle: Properties = {
+	color: 'rgba(254, 8, 16, 1)',
+	position: 'relative',
+	textAlign: 'center',
+	top: '8px',
+	padding: '4px',
+	fontFamily: 'Shlop',
+	fontSize: '40px',
+}
+
+export const formFieldStyle: Properties = {
+    padding: '8px',
+    width: '250px',
+    fontSize: '18px',
+    borderRadius: '4px',
+    border: '1px solid #fff',
+	marginBottom: '15px',
+	fontFamily: 'Shlop',
+}
+
+export const updateButtonStyle: Properties = {
+	backgroundColor: 'rgba(254, 8, 16, 1)',
+	position: 'relative',
+	height:'40px',
+	width:'80px',
+	fontFamily: 'Shlop',
+	fontSize: '14px',
+	alignSelf: 'center',
+	borderRadius: '6px',
+	border: 'none',
+	color:'white',
+	marginTop: '10px',
+	marginLeft: '20px',
+}
+
+//  export const successMessageStyle: Properties = {
+// 	color: 'rgba(254, 8, 16, 1)',
+// 	position: 'relative',
+// 	textAlign: 'center',
+// 	top: '8px',
+// 	padding: '4px',
+// 	fontFamily: 'Shlop',
+// 	fontSize: '12px',
+// }
+
+export const logoutButtonStyle: Properties = {
+	backgroundColor: 'rgba(254, 8, 16, 1)',
+	position: 'relative',
+	height:'40px',
+	width:'160px',
+	fontFamily: 'Shlop',
+	fontSize: '24px',
+	alignSelf: 'center',
+	borderRadius: '6px',
+	border: 'none',
+	color:'white',
+	marginBottom: '0px',
+}
